Update
- [x]  던전엔터 팝업 셋키 널 체크
- [x]  스킬 업그레이드 코스트 체크하기
- [x]  ~~저장타이밍~~
- ~~진화 성공시~~
- ~~보석 3000개 소모 또는 획득할때마다~~
- ~~전설, 유니크 유니온을 뽑았을 경우~~
- ~~60초마다 저장~~

+++ Assets/WOM/Scripts/Controller/GoldPigController.cs
@@ -158,7 +158,7 @@ public class GoldPigController : MonoBehaviour
     IEnumerator EnableGoldPigCor()
     {
 
-        var spawnGoldPigDelayTime = (float)Random.Range(StaticDefine.SPAWN_GOLDPIG_MIN_TIME, StaticDefine.SPAWN_GOLDPIG_MAX_TIME);
+        var spawnGoldPigDelayTime = Random.Range(StaticDefine.SPAWN_GOLDPIG_MIN_TIME, StaticDefine.SPAWN_GOLDPIG_MAX_TIME);
         
         var startTime = Time.time;
         float waitTime = spawnGoldPigDelayTime;
